fix(file-system): updated file-system code for images path in backend

+++ backend/src/filesystem/filesystem.ts
@@ -2,22 +2,39 @@ import path from "path";
 import fs from "fs/promises";
 import { randomUUID } from "crypto";
 
-const UPLOADS_ROOT = path.join(process.cwd(), "uploads");
+export const UPLOADS_ROOT = path.join(process.cwd(), "uploads");
 
-export async function saveFile(
+const extensionType: Record<string, string> = {
+  "image/jpeg": ".jpg",
+  "image/png": ".png",
+  "image/webp": ".webp",
+};
+
+export async function saveImage(
   buffer: Buffer,
-  originName: string,
-  subfolder: string,
+  mimeType: string,
+  subfolder: "listings",
 ): Promise<string> {
-  const ext = path.extname(originName);
+  const ext = extensionType;
+
+  if (!ext) {
+    throw new Error("Only JPEG, PNG and Webp images are allowed");
+  }
+
   const filename = `${randomUUID()}${ext}`;
   const dir = path.join(UPLOADS_ROOT, subfolder);
   await fs.mkdir(dir, { recursive: true });
   await fs.writeFile(path.join(dir, filename), buffer);
-  return path.join(subfolder, filename);
+  return path.posix.join(subfolder, filename);
 }
 
-export async function deleteFile(relativepath: string): Promise<void> {
-  const fullpath = path.join(UPLOADS_ROOT, relativepath);
+export async function deleteFile(relativePath: string): Promise<void> {
+  const normalizedpath = relativePath.replaceAll("\\", "/");
+
+  if (normalizedpath.startsWith("/") || normalizedpath.startsWith("..")) {
+    throw new Error("Invalid image path");
+  }
+
+  const fullpath = path.join(UPLOADS_ROOT, normalizedpath);
   await fs.rm(fullpath, { force: true });
 }
